Do not force refresh when write indexing buffer
